Move shared/modules/map_product_module.py to ssg module
 - New location is ssg/_products.py
 - Fixes all references to point to new location
