+++ ssg/__init__.py
@@ -2,6 +2,7 @@ import ssg._checks as checks
 import ssg._constants as constants
 import ssg._id_translate as id_translate
 import ssg._jinja as jinja
+import ssg._products as products
 import ssg._parse_oval as parse_oval
 import ssg._utils as utils
 import ssg._xccdf as xccdf
